@@ -2,10 +2,10 @@
   "name": "@nationallibraryofnorway/ngx-mime",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": ">=7.2.0 <8.0.0",
-    "@angular/core": ">=7.2.0 <8.0.0",
-    "@angular/flex-layout": ">=7.0.0 <8.0.0 || >=7.0.0-beta.23 <8.0.0",
-    "@angular/material": ">=7.2.1 <8.0.0",
+    "@angular/common": ">=8.0.0 <9.0.0",
+    "@angular/core": ">=8.0.0 <9.0.0",
+    "@angular/flex-layout": ">=8.0.0 <9.0.0 || >=8.0.0-beta.24 <8.0.0",
+    "@angular/material": ">=8.0.0 <9.0.0",
     "d3": ">=5.7.0 <6.0.0",
     "openseadragon": ">=2.4.0 <3.0.0"
   }
