@@ -1,6 +1,6 @@
 import { MimeDomHelper } from './../core/mime-dom-helper';
 import { ElementRef, Injectable } from '@angular/core';
-import { MatDialogConfig } from '@angular/material';
+import { MatDialogConfig } from '@angular/material/dialog';
 
 import { Dimensions } from './../core/models/dimensions';
 
