@@ -1,19 +1,17 @@
 import { NgModule } from '@angular/core';
-import {
-  MatToolbarModule,
-  MatButtonModule,
-  MatIconModule,
-  MatTooltipModule,
-  MatDialogModule,
-  MatTabsModule,
-  MatListModule,
-  MatSliderModule,
-  MatProgressSpinnerModule,
-  MatInputModule,
-  MatProgressBarModule,
-  MatCardModule,
-  MatSnackBarModule
-} from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatListModule } from '@angular/material/list';
+import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatSliderModule } from '@angular/material/slider';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatTabsModule } from '@angular/material/tabs';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatTooltipModule } from '@angular/material/tooltip';
 
 @NgModule({
   exports: [
